requirements.txt
python-telegram-bot[job-queue]==22.6
python-dotenv==1.2.1
Pillow>=10.0
google-generativeai>=0.8.0